+++ tianyancha/Pipfile
@@ -8,7 +8,6 @@ cpca = "*"
 requests = "*"
 lxml = "*"
 flask = "*"
-pyecharts = "*"
 
 [dev-packages]
