+++ tianyancha/f_cccs_scraper.py
@@ -6,8 +6,6 @@
  File: f_cccs_scraper.py
  Time: 7/24/18
 """
-import json
-import os
 import time
 
 import lxml
@@ -128,37 +126,5 @@ def main_scraper():
     print(_f_box)
 
 
-# ------------------------ data has been handle and forget above --------------------
-# 结构化后json数据文件
-
-TARGET_JSON = "structure_data.json"
-
-
-def generate_data_for_echarts():
-    if not os.path.exists('structure_data.json'):
-        return
-    cc_data = []
-    with open(TARGET_JSON, mode='r+', encoding='utf-8') as f_target_file:
-        json_map = json.load(fp=f_target_file)
-        for item in json_map:
-            if not item.get('county') == '':
-                location_name = item.get('county')
-            else:
-                if not item.get('city') == '':
-                    location_name = item.get('city')
-                else:
-                    location_name = item.get('province')
-            single_way = [{'name': '长春市'},
-                          {'name': location_name},
-                          {'name': item.get('title')},
-                          {'name': item.get('publish_time')},
-                          {'name': item.get('title_link')},
-                          {'name': item.get('raw_link')},
-                          {'name': item.get('location')}]
-            cc_data.append(single_way)
-    print(cc_data)
-
-
 if __name__ == '__main__':
-    # main_scraper()
-    generate_data_for_echarts()
+    main_scraper()
