+++ tianyancha/tianyancha_visual_data.htm
@@ -346,7 +346,7 @@ series.push(
 option = {
     backgroundColor: '#404a59',
     title : {
-        text: '长春长生',
+        text: '长春长生疫苗采购省份流向',
         subtext: '数据来源：天眼查&中国政府采购网',
         left: 'center',
         textStyle : {
